Fix for 64 bit
IntPtr is 32 bits or 64 bits, but a file descriptor is always 32 bits.

+++ RaspberryIRDotNet/FileSystem/Native.cs
@@ -40,7 +40,7 @@ namespace RaspberryIRDotNet.FileSystem
         [StructLayout(LayoutKind.Sequential)]
         public struct pollfd
         {
-            public IntPtr fd;
+            public int fd;
             public POLL_EVENTS events;
             public POLL_EVENTS revents;
         }
